Fix clippy and formatting
